Extracted select_task_based_on_progress()
